# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BalanceResponse {
  success: boolean;
  balance?: number;
  error?: string;
}

export interface PriceResponse {
  success: boolean;
  cost?: number | null;
  count: number;
  error?: string;
}

export interface BuyRequest {
  country: string;
  service?: string;
  maxPrice?: string;
  operator?: string;
}

export interface BuyResponse {
  success: boolean;
  orderId?: string;
  phone?: string;
  error?: string;
}

export interface StatusResponse {
  success: boolean;
  status?: string;
  error?: string;
}

export interface CancelResponse {
  success: boolean;
  result?: string;
  error?: string;
}

export type GetPricesParams = {
  service?: string;
  country: string;
};