# lib/api-zod/src/generated/types/balanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceResponse {
  success: boolean;
  balance?: number;
  error?: string;
}