# lib/api-zod/src/generated/types/buyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyRequest {
  country: string;
  service?: string;
  maxPrice?: string;
  operator?: string;
}