# lib/api-zod/src/generated/types/buyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyResponse {
  success: boolean;
  orderId?: string;
  phone?: string;
  error?: string;
}