# lib/api-zod/src/generated/types/priceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface PriceResponse {
  success: boolean;
  cost?: number | null;
  count: number;
  error?: string;
}