# lib/api-zod/src/generated/types/statusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hero SMS Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusResponse {
  success: boolean;
  status?: string;
  error?: string;
}